requirements.txt
Django>=4.1.13
djangorestframework
djongo
django-cors-headers
django-filter
djangorestframework-simplejwt
